pyproject.toml
[project]
name = "rp3-medicineagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface-hub>=1.3.1",
    "linkup-sdk>=0.9.0",
    "llama-cpp-python>=0.3.16",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
]
